* Updated vector containers to remove warnings when compiling for 64 bit modes in Visual Studio * Updated version to 1.6.8
git-svn-id: http://madara.googlecode.com/svn/trunk@1304 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/knowledge_engine/containers/Double_Vector.h
@@ -145,7 +145,7 @@ namespace Madara
          * @return the value of the entry. Modifications to this will
          *         not be reflected in the context. This is a local copy.
          **/
-        type operator[] (unsigned int index);
+        type operator[] (size_t index);
       
         /**
          * Sets a knowledge variable to a specified value
